requirement.txt
joblib==0.15.1
Django==3.0
Pillow==7.0.0
requests==2.23.0
numpy==1.18.1
pandas==1.0.3
tensorflow==2.1.0
annoy==1.16.3
dj-database-url==0.5.0
psycopg2==2.8.6